+++ src/types/type.ts
@@ -1,4 +1,5 @@
 import { CustomMarking, DotMarking } from 'react-native-calendars'
+import { LatLng } from 'react-native-maps'
 
 export type IStartEnd = {
   afterTimestamp: number
@@ -11,4 +12,12 @@ export type AccessToken = {
 
 export type CalendarEvent = {
   [date: string]: DotMarking & CustomMarking
+}
+
+export type RegionData = {
+  id: string
+  title: string
+  description?: string
+  createdAt: number
+  latLng: LatLng
 }
